+++ iceoryx_utils/include/iceoryx_utils/posix_wrapper/thread.hpp
@@ -23,7 +23,7 @@ namespace iox
 {
 namespace posix
 {
-constexpr uint8_t MAX_THREAD_NAME_LENGTH = 15U;
+constexpr size_t MAX_THREAD_NAME_LENGTH = 15U;
 
 using ThreadName_t = cxx::string<MAX_THREAD_NAME_LENGTH>;
